Added the abstract class for file out stream integration tests
